separate media streams into client and server concerns, move consumers and producers to network, webcam to client
