+++ packages/client-core/src/media/webcam/WebcamInputWorker.js
