+++ packages/client-core/src/media/webcam/faceEnvWorkerPatch.js
