+++ packages/client/src/pages/location/[locationName].tsx
@@ -9,8 +9,8 @@ import LoadLocationScene from '@xrengine/client-core/src/components/World/LoadLo
 import NetworkInstanceProvisioning from '@xrengine/client-core/src/components/World/NetworkInstanceProvisioning'
 import OfflineLocation from '@xrengine/client-core/src/components/World/OfflineLocation'
 import { LocationAction, useLocationState } from '@xrengine/client-core/src/social/services/LocationService'
-import { DefaultLocationSystems } from '@xrengine/client-core/src/systems/DefaultLocationSystems'
 import { AuthService } from '@xrengine/client-core/src/user/services/AuthService'
+import { DefaultLocationSystems } from '@xrengine/client-core/src/world/DefaultLocationSystems'
 import { SceneService } from '@xrengine/client-core/src/world/services/SceneService'
 import { Engine } from '@xrengine/engine/src/ecs/classes/Engine'
 import { useEngineState } from '@xrengine/engine/src/ecs/classes/EngineState'
